requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
torch==1.9.0
torchvision==0.10.0 -f https://download.pytorch.org/whl/torch_stable.html
Pillow